+++ domain/src/main/java/th/or/nectec/tanrabad/domain/geographic/FilterBoundaryCalculator.java
@@ -20,6 +20,6 @@ package th.or.nectec.tanrabad.domain.geographic;
 import th.or.nectec.tanrabad.entity.Location;
 
 public interface FilterBoundaryCalculator {
-    Location getMinLocation(Location currentLocation, int distanceInKm);
-    Location getMaxLocation(Location currentLocation, int distanceInKm);
+    Location getMinLocation(Location currentLocation, double distanceInKm);
+    Location getMaxLocation(Location currentLocation, double distanceInKm);
 }
